+++ test/test_helper.exs
@@ -1,1 +1,2 @@
+{:ok, _} = Registry.start_link(keys: :unique, name: Dstar.Utility.StreamRegistry)
 ExUnit.start()